+++ src/server/Balance/BalanceRoute.ts
@@ -10,7 +10,7 @@ router.get("/", (req, res) => {
     request: req,
     response: res
   };
-  server.list(node);
+  server.listRequestHandler(node);
 });
 
 router.post("/rebuild", (req, res) => {